+++ src/data_gen/data_loader.hpp
@@ -16,5 +16,5 @@ extern "C"
     void get_indices(const char* file_name, Size* indices);
     Size get_num_positions(const Size* indices, const Size* selection, Size n_selected);
     Size get_dense_psq_data(const char* file_name, const Size* selection, Size n_selected, char* X, short* evals, char* results);
-    Size get_sparse_psq_data(const char* file_name, const Size* selection, Size n_selected, Size* idx, unsigned short* cols, char* colors, short* evals, char* results);
+    Size get_sparse_psq_data(const char* file_name, const Size* selection, Size n_selected, Size* idx, unsigned int* cols, char* colors, short* evals, char* results);
 }
